Add HI-style pu subtraction.
This is the first draft of the Heavy Ion background subtraction.

The background subtraction is eta-dependent. For each gct eta value, the average region et is found. Currently this means 22 different average pu levels, one for each eta index. During jet finding, each region has the appropriate background amount subtracted (with a floor of 0) before looking for jets. Because of the nature of algorithm, "uncorrected" jets are meaningless, so the Jet and corrJet collections are identical in this case.

Care was taken to avoid making any changes to non-HI logic. Behavior with Taus and EG candidates is currently unknown/untested when using the HI algorithm.